+++ packages/charts/react-charts/library/src/components/CommonComponents/useCartesianChartStyles.styles.ts
@@ -2,7 +2,7 @@ import { makeStyles, mergeClasses } from '@griffel/react';
 import { CartesianChartProps, CartesianChartStyles } from './CartesianChart.types';
 import { SlotClassNames } from '@fluentui/react-utilities/src/index';
 import { tokens, typographyStyles } from '@fluentui/react-theme';
-import { useRtl } from '../../utilities/utilities';
+import { HighContrastSelector, useRtl } from '../../utilities/utilities';
 
 /**
  * @internal
@@ -15,6 +15,7 @@ export const cartesianchartClassNames: SlotClassNames<CartesianChartStyles> = {
   yAxis: 'fui-cart__yAxis',
   opacityChangeOnHover: 'fui-cart__opacityChangeOnHover',
   legendContainer: 'fui-cart__legendContainer',
+  svgTooltip: 'fui-cart_svgTooltip',
   shapeStyles: 'fui-cart__shapeStyles',
   descriptionMessage: 'fui-cart__descriptionMessage',
   hover: 'fui-cart__hover',
@@ -43,6 +44,9 @@ const useStyles = makeStyles({
     textAlign: 'center',
     color: tokens.colorNeutralForeground2,
     fill: tokens.colorNeutralForeground1,
+    [HighContrastSelector]: {
+      fill: 'CanvasText',
+    },
   },
   xAxis: {
     '& text': {
@@ -89,6 +93,12 @@ const useStyles = makeStyles({
     marginTop: tokens.spacingVerticalS,
     marginLeft: tokens.spacingHorizontalXL,
   },
+  svgTooltip: {
+    fill: tokens.colorNeutralBackground1,
+    [HighContrastSelector]: {
+      fill: 'Canvas',
+    },
+  },
 });
 /**
  * Apply styling to the Carousel slots based on the state
@@ -117,5 +127,6 @@ export const useCartesianChartStyles = (props: CartesianChartProps): CartesianCh
       cartesianchartClassNames.legendContainer,
       baseStyles.legendContainer /*props.styles?.legendContainer*/,
     ),
+    svgTooltip: mergeClasses(cartesianchartClassNames.svgTooltip, baseStyles.svgTooltip /*props.styles?.svgTooltip*/),
   };
 };