+++ packages/charts/react-charts/library/src/components/HorizontalBarChartWithAxis/useHorizontalBarChartWithAxisStyles.styles.ts
@@ -23,6 +23,7 @@ export const hbcWithAxisClassNames: SlotClassNames<HorizontalBarChartWithAxisSty
   chartTitle: '',
   shapeStyles: '',
   chartWrapper: '',
+  svgTooltip: '',
 };
 
 const useStyles = makeStyles({