+++ packages/charts/react-charts/library/src/components/ScatterChart/useScatterChartStyles.styles.ts
@@ -19,6 +19,7 @@ export const scatterChartClassNames: SlotClassNames<ScatterChartStyles> = {
   opacityChangeOnHover: 'fui-line__opacityChangeOnHover',
   shapeStyles: 'fui-line__shapeStyles',
   chartWrapper: 'fui-line__chartWrapper',
+  svgTooltip: '',
 };
 
 /**