+++ packages/react-examples/src/react-charting/HorizontalBarChartWithAxis/HorizontalBarChartWithAxis.doc.tsx
@@ -7,6 +7,7 @@ import { HorizontalBarChartWithAxisTooltipExample } from './HorizontalBarChartWi
 import { HorizontalBarChartWithAxisStringAxisTooltipExample } from './HorizontalBarChartWithAxis.StringAxisTooltip.Example';
 import { HorizontalBarChartWithAxisDynamicExample } from './HorizontalBarChartWithAxis.Dynamic.Example';
 import { HorizontalBarChartWithAxisNegativeExample } from './HorizontalBarChartWithAxis.Negative.Example';
+import { HBWAAxisCategoryOrderExample } from './HorizontalBarChartWithAxis.AxisCategoryOrder.Example';
 
 const HorizontalBarChartWithAxisBasicExampleCode =
   require('!raw-loader?esModule=false!@fluentui/react-examples/src/react-charting/HorizontalBarChartWithAxis/HorizontalBarChartWithAxis.Basic.Example.tsx') as string;
@@ -18,6 +19,8 @@ const HorizontalBarChartWithAxisDynamicExampleCode =
   require('!raw-loader?esModule=false!@fluentui/react-examples/src/react-charting/HorizontalBarChartWithAxis/HorizontalBarChartWithAxis.Dynamic.Example.tsx') as string;
 const HorizontalBarChartWithAxisNegativeExampleCode =
   require('!raw-loader?esModule=false!@fluentui/react-examples/src/react-charting/HorizontalBarChartWithAxis/HorizontalBarChartWithAxis.Negative.Example.tsx') as string;
+const HBWAAxisCategoryOrderExampleCode =
+  require('!raw-loader?esModule=false!@fluentui/react-examples/src/react-charting/HorizontalBarChartWithAxis/HorizontalBarChartWithAxis.AxisCategoryOrder.Example.tsx') as string;
 
 export const HorizontalBarChartWithAxisPageProps: IDocPageProps = {
   title: 'HorizontalBarChartWithAxis',
@@ -49,6 +52,11 @@ export const HorizontalBarChartWithAxisPageProps: IDocPageProps = {
       title: 'HorizontalBarChartWithAxis negative X axis',
       code: HorizontalBarChartWithAxisNegativeExampleCode,
       view: <HorizontalBarChartWithAxisNegativeExample />,
+    },
+    {
+      title: 'HorizontalBarChartWithAxis Axis Category Order',
+      code: HBWAAxisCategoryOrderExampleCode,
+      view: <HBWAAxisCategoryOrderExample />,
     },
   ],
   overview: require<string>('!raw-loader?esModule=false!@fluentui/react-examples/src/react-charting/HorizontalBarChartWithAxis/docs/HorizontalBarChartWithAxisOverview.md'),